bills migration tipe and customer autocomplete fix

+++ database/migrations/2022_10_06_184034_bills_update_type_of_fields.php
@@ -13,7 +13,7 @@ return new class extends Migration {
     public function up()
     {
         Schema::table('bills', function (Blueprint $table) {
-            $table->unsignedBigInteger('type')->change();
+            $table->integer('type')->change();
             $table->unsignedBigInteger('bill_number')->change();
             $table->unsignedBigInteger('currency')->change();
             $table->unsignedBigInteger('exchange')->change();
